Day 7
Add code for managing directories and files recursively. Implement functions to add directories and files, calculate total weight, and print the directory tree. Also, add a function to find the size of the smallest directory that meets a given minimum size.

In 7_2.cpp, calculate the required space for an update on Drive C and find the smallest file that can be removed to make enough space.